requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
bestchange-api==2.0.0.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
dateparser==1.1.0
Deprecated==1.2.13
filelock==3.4.2
Flask==2.0.2
Flask-Cors==3.0.10
frozenlist==1.2.0
grpcio==1.43.0
idna==3.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.3.3
MarkupSafe==2.0.1
msgpack==1.0.3
multidict==5.2.0
numpy==1.22.0
packaging==21.3
pandas==1.3.5
protobuf==3.19.1
pyparsing==3.0.6
pyrsistent==0.18.0
python-binance==1.0.15
python-dateutil==2.8.2
python-dotenv==0.19.2
pythonpancakes==1.0.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
ray==1.9.1
redis==4.1.0
regex==2021.11.10
requests==2.27.0
six==1.16.0
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
urllib3==1.26.7
websockets==9.1
Werkzeug==2.0.2
wrapt==1.13.3
yarl==1.7.2
zipp==3.7.0
